requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-colorfield==0.1.14
django-countries==5.3
django-localflavor==2.0
django-widget-tweaks==1.4.2
idna==2.6
oauthlib==2.0.7
Pillow==5.1.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
